+++ src/app/ClientSideInit.tsx
@@ -3,15 +3,14 @@
 // For applying global client-side effects like auth state monitoring and background music
 
 import type { User } from "firebase/auth";
-import { useEffect } from "react";
+import { useEffect, useRef } from "react";
 import { toast } from "react-toastify";
 // Firebase module
 import { onAuthStateChangedListener } from "@/services/firebase";
-import { getWallet } from "@/services/game-apis";
-import { useCoins, useUser, useXP } from "@/services/store";
+import { getWallet, signIn } from "@/services/game-apis";
+import { useCoins, useProfile, useUser, useXP } from "@/services/store";
 
 const ClientSideInit = (): null => {
-	const user = useUser((state) => state.user);
 	const setUser = useUser(
 		(state): ((newUser: User | null) => void) => state.setUser,
 	);
@@ -22,26 +21,48 @@ const ClientSideInit = (): null => {
 		(state): ((newCoins: number) => void) => state.setCoins,
 	);
 	const setXP = useXP((state): ((newXP: number) => void) => state.setXP);
+	const setName = useProfile((state) => state.setName);
+	const setEmail = useProfile((state) => state.setEmail);
+	const setPic = useProfile((state) => state.setPic);
 
-	// Load user
-	useEffect((): (() => void) => {
-		const unsubscribe = onAuthStateChangedListener(
-			async (usr): Promise<void> => {
-				setUser(usr);
-				if (!usr) {
-					setAuthReady(false);
-				}
-			},
-		);
-		return (): void => unsubscribe();
-	}, [setUser, setAuthReady]);
+	// Track pending sign-in attempts
+	const pendingSignInRef = useRef<Map<string, boolean>>(new Map());
+
+	// Handle user authentication and backend sign-in
 	useEffect(() => {
-		if (user != null) {
-			const fetchWallet = async () => {
-				try {
-					const token = await user.getIdToken();
-					const wallet = await getWallet(token);
+		const unsubscribe = onAuthStateChangedListener(
+			async (firebaseUser): Promise<void> => {
+				if (!firebaseUser) {
+					// User signed out
+					setUser(null);
+					setAuthReady(false);
+					return;
+				}
 
+				// Check if we've already processed this user
+				const userId = firebaseUser.uid;
+				if (pendingSignInRef.current.get(userId)) {
+					return; // Already processing
+				}
+
+				try {
+					pendingSignInRef.current.set(userId, true);
+
+					// Step 1: Get Firebase token
+					const idToken = await firebaseUser.getIdToken();
+
+					// Step 2: Sign in to backend
+					const backendUser = await signIn(idToken);
+					console.log("Backend user data:", backendUser);
+
+					// Step 3: Update user state
+					setUser(firebaseUser);
+					setName(backendUser.name);
+					setEmail(backendUser.email);
+					setPic(backendUser.profile_pic);
+
+					// Step 4: Get wallet
+					const wallet = await getWallet(idToken);
 					if (wallet.success) {
 						setCoins(wallet.coins);
 						setXP(wallet.xp);
@@ -50,18 +71,18 @@ const ClientSideInit = (): null => {
 						toast.error("get wallet failed");
 						setAuthReady(false);
 					}
-				} catch (err) {
-					console.error("Error fetching wallet:", err);
-					toast.error("Something went wrong while fetching wallet");
+				} catch (error) {
+					console.error("Auth error:", error);
+					toast.error("Authentication failed");
 					setAuthReady(false);
+				} finally {
+					pendingSignInRef.current.delete(userId);
 				}
-			};
+			},
+		);
 
-			fetchWallet();
-		} else {
-			setAuthReady(false);
-		}
-	}, [user, setCoins, setXP, setAuthReady]);
+		return (): void => unsubscribe();
+	}, [setUser, setAuthReady, setCoins, setXP, setName, setEmail, setPic]);
 
 	return null;
 };
