+++ src/app/Menu.tsx
@@ -14,16 +14,11 @@ import ShortcutModal from "@/modals/ShortcutModal";
 import SoundConfigModal from "@/modals/SoundConfigModal";
 import TutorialModal from "@/modals/TutorialModal";
 import { signInWithGoogle, signOutUser } from "@/services/firebase";
-import { signIn } from "@/services/game-apis";
-import { useProfile, useUser } from "@/services/store";
+import { useUser } from "@/services/store";
 import type { MenuModalType } from "@/services/types";
 
 const Menu = () => {
 	const user = useUser((state) => state.user);
-	const setUser = useUser((state) => state.setUser);
-	const setName = useProfile((state) => state.setName);
-	const setEmail = useProfile((state) => state.setEmail);
-	const setPic = useProfile((state) => state.setPic);
 	const router = useRouter();
 	const { canShowToast, resetCooldown } = useToastCooldown(TOAST_DURATION);
 	const [activeModal, setActiveModal] = useState<MenuModalType>(null);
@@ -40,24 +35,8 @@ const Menu = () => {
 
 	const handleSignIn = async () => {
 		try {
-			// Step 1: Firebase popup
-			const user = await signInWithGoogle();
-			if (!user) throw new Error("No user returned from Google Sign-In");
+			await signInWithGoogle();
 
-			// Step 2: Get Firebase ID token
-			const idToken = await user.getIdToken();
-
-			// Step 3: Call backend sign-in API
-			const backendUser = await signIn(idToken);
-			// TODO: Use these values in the app as needed and delete these console logs
-			console.log("Backend user data:", backendUser);
-			console.log("Is New Account:", backendUser.new_account); // returns true if new account
-			// Step 4: Update global user state (TODO)
-			setUser(user);
-			setName(backendUser.name);
-			setEmail(backendUser.email);
-			setPic(backendUser.profile_pic);
-			// Step 5: Dismiss any existing sign-in error toasts
 			toast.dismiss(TOAST_IDS.User.SignInError);
 			resetCooldown();
 		} catch (error) {
@@ -68,7 +47,6 @@ const Menu = () => {
 	const handleSignOut = async () => {
 		try {
 			await signOutUser();
-			setUser(null);
 		} catch (error) {
 			console.error("Sign out error:", error);
 		}
